refactor(kubernetes): introduce resources helpers in kubernetes
